# HAJIMI_UI/docs/api-contract-demo.yaml
openapi: 3.0.3
info:
  title: HAJIMI Demo API Contract
  description: |
    HAJIMI 智能桌面指引助手 Demo 阶段前后端接口契约。
    本契约覆盖「截图 → AI 识图 → 输出操作步骤 → 屏幕标注」核心闭环，
    供 A（后端/AI）与 B（前端/PyQt5）在 Demo 开发阶段对齐使用。
  version: 1.0.0
  contact:
    name: HAJIMI Team
servers:
  - url: http://localhost:8000
    description: 本地开发环境（FastAPI 默认端口）

# Demo 阶段统一认证：固定 Demo Key，生产阶段替换为 JWT/HMAC
security:
  - DemoKeyAuth: []

components:
  securitySchemes:
    DemoKeyAuth:
      type: apiKey
      in: header
      name: X-Demo-Key
      description: Demo 阶段固定值为 `hajimi-demo-2026`

  schemas:
    # ────────────────────────── 请求/响应模型 ──────────────────────────
    ProcessRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: 用户自然语言提问
          example: "怎么安装微信？"
        image:
          type: string
          format: byte
          description: |
            屏幕截图 Base64 编码（建议 JPEG/PNG，分辨率不超过 1920×1080，
            单张大小控制在 2MB 以内）。
            Demo 阶段可传空，后端返回预置 Mock 数据。
          example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
        window_title:
          type: string
          description: 当前活动窗口标题（用于意图理解上下文）
          example: "桌面"
        context:
          type: array
          description: 多轮对话上下文（最近 3 轮）
          items:
            $ref: "#/components/schemas/ChatTurn"
          maxItems: 3

    ProcessResponse:
      type: object
      required:
        - task_id
        - success
        - intent
        - blueprint
        - steps
      properties:
        task_id:
          type: string
          format: uuid
          description: 本次任务唯一 ID，后续 step/clarify/report 均需携带
          example: "550e8400-e29b-41d4-a716-446655440000"
        success:
          type: boolean
          description: 处理是否成功
          example: true
        intent:
          $ref: "#/components/schemas/Intent"
        ui_elements:
          type: array
          description: 截图中识别到的 UI 元素列表
          items:
            $ref: "#/components/schemas/UIElement"
        annotated_image:
          type: string
          format: byte
          description: 带 SoM 编号标注的截图 Base64（前端可直接显示）
          example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
        blueprint:
          $ref: "#/components/schemas/Blueprint"
        steps:
          type: array
          description: 当前任务的操作步骤列表
          items:
            $ref: "#/components/schemas/Step"
        reference_resolution:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
          description: 截图参考分辨率 [width, height]，bbox 坐标系与此一致
          example: [2560, 1440]
        detection_meta:
          type: object
          description: 检测元数据（latency_ms, element_count, backend）
          additionalProperties: true

    InspectRequest:
      type: object
      required:
        - image
      properties:
        image:
          type: string
          description: Base64 截图或 data URI
        screen_width:
          type: integer
          minimum: 1
        screen_height:
          type: integer
          minimum: 1

    InspectResponse:
      type: object
      required:
        - success
        - ui_elements
        - reference_resolution
      properties:
        success:
          type: boolean
        ui_elements:
          type: array
          items:
            $ref: "#/components/schemas/UIElement"
        annotated_image:
          type: string
          description: SoM 标注图 Base64
        reference_resolution:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
        detection_meta:
          type: object
          additionalProperties: true

    StepRequest:
      type: object
      required:
        - task_id
        - action
      properties:
        task_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        action:
          type: string
          enum: [advance, rollback, skip, terminate]
          description: |
            - advance: 用户已完成当前步骤，请求下一步
            - rollback: 用户要求回退一步
            - skip: 用户要求跳过当前步骤
            - terminate: 用户终止任务
          example: "advance"
        step_index:
          type: integer
          description: 当前步骤序号（用于校验）
          example: 1
        fingerprint:
          type: string
          description: 当前屏幕指纹 SHA256 值，用于蓝图状态机校验
          example: "a1b2c3d4e5f6..."

    StepResponse:
      type: object
      required:
        - task_id
        - action
        - current_step
        - blueprint_state
      properties:
        task_id:
          type: string
          format: uuid
        action:
          type: string
          enum: [advance, rollback, skip, suspended, complete, terminated]
          description: 后端决定的状态推进结果
        current_step:
          type: integer
          description: 当前步骤指针（1-based）
          example: 2
        next_step:
          $ref: "#/components/schemas/Step"
        blueprint_state:
          type: string
          enum: [generated, pending_confirm, executing, suspended, rolling_back, completed, terminated]
        message:
          type: string
          description: 当状态为 suspended 时，返回给用户的提示文案
          example: "检测到屏幕状态与预期不符，您要跳过此步还是回退重试？"

    ClarifyRequest:
      type: object
      required:
        - task_id
        - answer
      properties:
        task_id:
          type: string
          format: uuid
        answer:
          type: string
          description: 用户对澄清问题的回答
          example: "我想保存当前打开的 Word 文档"

    ClarifyResponse:
      type: object
      required:
        - task_id
        - confidence
        - needs_clarification
      properties:
        task_id:
          type: string
          format: uuid
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.93
        needs_clarification:
          type: boolean
          example: false
        question:
          type: string
          nullable: true
          description: 仍不够清晰时继续提问
          example: null
        updated_intent:
          $ref: "#/components/schemas/Intent"

    ReportRequest:
      type: object
      required:
        - task_id
      properties:
        task_id:
          type: string
          format: uuid
        result:
          type: string
          enum: [success, fail, cancel, redirect]
          example: "success"
        feedback_type:
          type: string
          enum: [useful, useless, neutral]
          example: "useful"
        duration_ms:
          type: integer
          description: 任务总耗时（毫秒）
          example: 5200
        comment:
          type: string
          description: 用户评语（可选）
          example: "指引很清晰"

    ReportResponse:
      type: object
      required:
        - received
      properties:
        received:
          type: boolean
          example: true

    # ────────────────────────── 业务数据模型 ──────────────────────────
    ChatTurn:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string

    Intent:
      type: object
      required:
        - category
        - summary
        - confidence
        - needs_clarification
      properties:
        category:
          type: string
          enum:
            - operation_guide
            - element_cognition
            - error_diagnosis
            - ui_navigation
            - content_cognition
            - file_management
            - proactive_alert
            - tutorial_generation
            - emotion_comfort
          description: 九大意图域之一
          example: "operation_guide"
        summary:
          type: string
          description: 脱敏后的意图摘要
          example: "安装微信"
        reference_type:
          type: string
          enum: [explicit, visual, deictic, fuzzy, context]
          description: 五种指代方式之一
          example: "explicit"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.92
        needs_clarification:
          type: boolean
          description: 置信度 < 0.8 时为 true
          example: false

    UIElement:
      type: object
      required:
        - element_id
        - bbox
        - element_type
        - confidence
      properties:
        element_id:
          type: string
          description: SoM 全局唯一编号
          example: "~1"
        bbox:
          type: array
          items:
            type: integer
          minItems: 4
          maxItems: 4
          description: 边界框 [x1, y1, x2, y2]，坐标系为截图左上角原点
          example: [120, 340, 240, 380]
        element_type:
          type: string
          enum: [button, input, icon, menu, checkbox, dropdown, text, other]
          example: "icon"
        text:
          type: string
          description: 元素上的文字（OCR 识别结果，可能为空）
          example: "Microsoft Edge"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.95
        center:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
          description: 元素中心点坐标 [cx, cy]
          example: [180, 360]

    Step:
      type: object
      required:
        - step_index
        - action
        - description
        - status
      properties:
        step_index:
          type: integer
          description: 步骤序号，从 1 开始
          example: 1
        action:
          type: string
          description: 简短动作文案
          example: "打开浏览器"
        description:
          type: string
          description: 给用户看的详细说明
          example: "找到桌面上的 Microsoft Edge 图标，双击打开浏览器。"
        target_element_id:
          type: string
          nullable: true
          description: 目标元素 SoM 编号
          example: "~1"
        status:
          type: string
          enum: [pending, active, done, skipped, failed]
          description: 步骤状态
          example: "active"
        annotation:
          $ref: "#/components/schemas/Annotation"

    Annotation:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [arrow_highlight, highlight_only, arrow_only, label_only, none]
          description: 标注类型
          example: "arrow_highlight"
        arrow_from:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
          description: 箭头起点 [x, y]
          example: [50, 360]
        arrow_to:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
          description: 箭头终点 [x, y]，建议指向目标元素中心
          example: [180, 360]
        highlight_bbox:
          type: array
          items:
            type: integer
          minItems: 4
          maxItems: 4
          description: 高亮框 [x1, y1, x2, y2]
          example: [120, 340, 240, 380]
        label_position:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
          description: 编号标签左上角坐标 [x, y]
          example: [120, 296]
        label_text:
          type: string
          description: 标签显示文字
          example: "~1"

    Blueprint:
      type: object
      required:
        - name
        - total_steps
        - current_step
        - state
      properties:
        name:
          type: string
          description: 蓝图名称
          example: "安装微信"
        total_steps:
          type: integer
          example: 3
        current_step:
          type: integer
          example: 1
        state:
          type: string
          enum: [generated, pending_confirm, executing, suspended, rolling_back, completed, terminated]
          example: "pending_confirm"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "缺少必填字段 query"
            details:
              type: object
              additionalProperties: true

  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "INVALID_REQUEST"
              message: "缺少必填字段 query"
    Unauthorized:
      description: 认证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "AUTH_FAILED"
              message: "X-Demo-Key 无效"
    NotFound:
      description: 任务不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "NOT_FOUND"
              message: "task_id 不存在"
    InternalError:
      description: 服务端内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "LLM 调用超时"

paths:
  # ────────────────────────── 核心接口 ──────────────────────────
  /api/demo/process:
    post:
      summary: 核心流程入口：识图 + 理解 + 生成步骤与标注
      description: |
        前端上传截图与用户问题，后端完成 UI 解析、SoM 标注、意图理解、
        蓝图生成，返回带标注坐标的操作步骤。
      tags:
        - Demo Core
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRequest"
            example:
              query: "怎么安装微信？"
              image: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
              window_title: "桌面"
              context: []
      responses:
        "200":
          description: 成功返回任务蓝图与步骤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResponse"
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                success: true
                intent:
                  category: "operation_guide"
                  summary: "安装微信"
                  reference_type: "explicit"
                  confidence: 0.92
                  needs_clarification: false
                ui_elements:
                  - element_id: "~1"
                    bbox: [120, 340, 240, 380]
                    element_type: "icon"
                    text: "Microsoft Edge"
                    confidence: 0.95
                    center: [180, 360]
                  - element_id: "~2"
                    bbox: [860, 620, 1020, 660]
                    element_type: "button"
                    text: "下载"
                    confidence: 0.91
                    center: [940, 640]
                annotated_image: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                blueprint:
                  name: "安装微信"
                  total_steps: 3
                  current_step: 1
                  state: "pending_confirm"
                steps:
                  - step_index: 1
                    action: "打开浏览器"
                    description: "找到桌面上的 Microsoft Edge 图标，双击打开浏览器。"
                    target_element_id: "~1"
                    status: "active"
                    annotation:
                      type: "arrow_highlight"
                      arrow_from: [50, 360]
                      arrow_to: [180, 360]
                      highlight_bbox: [120, 340, 240, 380]
                      label_position: [120, 296]
                      label_text: "~1"
                  - step_index: 2
                    action: "访问微信官网"
                    description: "在浏览器地址栏输入 weixin.qq.com 并回车。"
                    target_element_id: "~2"
                    status: "pending"
                    annotation:
                      type: "highlight_only"
                      highlight_bbox: [860, 620, 1020, 660]
                  - step_index: 3
                    action: "点击下载按钮"
                    description: "在官网首页找到「下载」按钮并点击。"
                    target_element_id: "~2"
                    status: "pending"
                    annotation:
                      type: "arrow_highlight"
                      arrow_from: [50, 640]
                      arrow_to: [940, 640]
                      highlight_bbox: [860, 620, 1020, 660]
                      label_position: [860, 576]
                      label_text: "~2"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/demo/inspect:
    post:
      summary: 元素检测检验（无步骤生成）
      description: |
        仅对截图运行 UI 检测，返回全量 ui_elements 与 SoM 标注图，供 B 端检验模式验收。
      tags:
        - Demo Core
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectRequest"
      responses:
        "200":
          description: 检测成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: 未检测到 UI 元素
        "502":
          description: 检测器调用失败

  /api/demo/step:
    post:
      summary: 推进/回退/跳过/终止蓝图步骤
      description: |
        用户执行完一步后，前端调用此接口推进蓝图。
        后端根据当前屏幕指纹校验状态，决定是推进、挂起还是完成。
      tags:
        - Demo Core
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepRequest"
            example:
              task_id: "550e8400-e29b-41d4-a716-446655440000"
              action: "advance"
              step_index: 1
              fingerprint: "a1b2c3d4e5f6789012345678901234567890abcdef"
      responses:
        "200":
          description: 成功推进步骤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepResponse"
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                action: "advance"
                current_step: 2
                blueprint_state: "executing"
                next_step:
                  step_index: 2
                  action: "访问微信官网"
                  description: "在浏览器地址栏输入 weixin.qq.com 并回车。"
                  target_element_id: "~2"
                  status: "active"
                  annotation:
                    type: "highlight_only"
                    highlight_bbox: [860, 620, 1020, 660]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/demo/clarify:
    post:
      summary: 主动澄清应答
      description: |
        当 process 返回 needs_clarification=true 时，前端展示澄清问题，
        用户回答后调用此接口更新意图。
      tags:
        - Demo Core
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClarifyRequest"
            example:
              task_id: "550e8400-e29b-41d4-a716-446655440000"
              answer: "我想保存当前打开的 Word 文档"
      responses:
        "200":
          description: 返回更新后的意图与置信度
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClarifyResponse"
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                confidence: 0.93
                needs_clarification: false
                question: null
                updated_intent:
                  category: "operation_guide"
                  summary: "保存 Word 文档"
                  reference_type: "context"
                  confidence: 0.93
                  needs_clarification: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/demo/report:
    post:
      summary: 审计与反馈上报
      description: |
        任务结束后（成功/失败/取消），前端异步上报结果与用户反馈。
        Demo 阶段后端仅记录到内存/日志，不写入数据库。
      tags:
        - Demo Audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
            example:
              task_id: "550e8400-e29b-41d4-a716-446655440000"
              result: "success"
              feedback_type: "useful"
              duration_ms: 5200
              comment: "指引很清晰"
      responses:
        "200":
          description: 上报成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
              example:
                received: true
        "400":
          $ref: "#/components/responses/BadRequest"

  # ────────────────────────── 健康检查 ──────────────────────────
  /api/demo/health:
    get:
      summary: 服务健康检查
      description: 供前端启动时探测后端是否可用。
      tags:
        - Demo System
      security: []
      responses:
        "200":
          description: 服务正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  version:
                    type: string
                    example: "1.0.0"